backend/photo_blog/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-environ==0.13.0
psycopg[binary]==3.3.3
boto3==1.42.66
pillow==12.1.1
ExifRead==3.5.1
resend==2.26.0
requests==2.32.5
gunicorn==23.0.0
